Split backfill_db.py functions based on purpose.
The backfill_db.py file was beginning to become a bit of a hassle to
scroll around in. More importantly, however, is the fact that the
filename doesn't corrlate to the functions in said file. backfill_db.py
is meant to backfill the database, and as such should only contain
functions that work with the database in some capacity. Thus, the
URL-related functions and the reddit-related functions each now have
their own separate file.