Change conditional to avoid short-circuit eval
Using the or statement resulted in the potential that even though an
Imgur link is not a single image, it not also being an invalid link
would result in the wrong processing. Also, this commit changes the way
in which a single-image (one ending in .jpeg, ,jpg, or .png).
Previously, it was concatenated to a potentially empty list. This
resulted in exploding the outfit URL into single characters. The new
method is to simply use the append() method of the list object, to avoid
such explosions.