Add check if Imgur JSON is one or multple objects
When querying the API on the /image endpoint, it returns a single JSON
object (as in, it is not in a list). This differs from the /gallery or
/album endpoints, which do return JSON objects in lists. This commit
also reflects name changes, namely that of image_hash to imgur_hash.
Also used a list comprehension to filter URLs for only ones that are
Imgur or Dressed.so. As well as that, this commit changes the way in
which a comment is split when extracting images in plaintext.
Previously, it split on spaces, which resulted in keeping carriage
returns (newlines, or \n) in tokens. The goal was in fact to split on
newlines as well. As such, the new way in which a comment is split for
plaintext image URL extraction is by simply calling .split(), which
splits on ALL whitespace.