Add more specific checks for dressed.so links
This commit also reverts a change in a previous commit. Namely that of
splitting all URLs and reformatting them with the .png attachment. Due
to the nature of extracting URLs from Imgur albums and galleries, which
may or may not end in .png, for the time being, URLs are left as is, as
necessary. That is to say, i.redd.it URLs will keep their original file
extension, as well as cdn.dressed.so URLs.