Revert assignment of Imgur URL information
This commit reverts the way in which information about an Imgur URL is
assigned. Previously, it used multiple assignments. However, because the
output of create_imgur_url_info is a dictionary, this does not work, and
simply just assigns the keys to the variables on the left hand side of
the equal sign. As such, this commit reverts back to the original
assignemnt decision, that of simply having two assignments.