Change thread_id to be alphanumeric hash instead
Reddit submissions have two IDs, one that, at its core, has a hyphen,
and another, which does not. Both are alphanumeric otherwise. The 'id'
field of the Submission object from the Python Reddit API Wrapper refers
to the unhyphenated ID, while the 'link_id' of the Comment object from
the PRAW refers to the hyphenated ID. Thus, to ensure the IDs match,
this commit changes the 'thread_id' field of the comment dictionary we
create to be the alphanumeric unhyphenated ID found in the Submission
object.