Refactor comment extraction for dead URLs/comments
This commit primarily focuses on refactoring the way in which comments
are treated and processed when either an author/user is deleted or the
image URLs in a comment are invalid. Specifically, if the author is
deleted, it has no Author object (from the PRAW library), so the comment
does not have an author. Thus, we claim the author to be '[deleted]' and
store the rest of the information. This commit also tidies up the way in
which Imgur URLs are treated. Specifically, it calls
extract_image_urls_from_imgur_url in any case of an Imgur URL, to
prevent dead or unviewable URLs from being added. In addition, this
commit renames the loop variable of the main for loop in
create_outfit_urls, as well as tidy up some comments.