Refactor conditional for imgur.com/{has}
This commit refactors the way in which we check if an Imgur URL is an
Imgur image (that is, an Imgur URL that is of the form
http[s]://imgur.com/{hash}/). We now check if the path is simply /. If
it is, we know that the URL is http[s]://imgur.com/, and we ignore it.
Else, we process it the same way as before. This commit also changes the
information regarding an Imgur URL. Specifically, for a single image
(one ending in .jpg, .jpeg, or .png), a single image's 'url_type' is now
treated as 'image'. This is because on the chance the URL is dead, the
hash of the potentially dead image will be queried against the Imgur
API, and will be processed only if the URL is indeed not dead.