Remove date/time posted fields and add timestamp
Using the string version of the Unix time stamp returned from a comment
and thread isn't best when checking if a thread or comment is more than
a certain time interval old. So, comments and threads now have the
timestamp stored to represent their age, and we will convert the
timestamp on the front-end to a displayable time/date for the user.