Add file defining database class
This commit condenses any function that makes a query or has any
interaction with the database into one file, in one class. The reason
for this is because there would otherwise be many different cursors
acting on the database at once. This format drastically lowers the
amount of cursors that will be active on the database at any given
moment.