Remove tests for comment extraction
Previously there were tests that tested the comment extraction function.
However, the comment extraction function now tests if the URL is up
before adding it to the list of outfit URLs to be returned. Thus,
because any given URL could be down at any given moment, it's no longer
easy/reliable to test the extraction function. However, in prior
iterations of this function, comment extraction was successful.