Change structure of displaying list of threads
Previously there were only two components for displaying thread
information: one component (Thread) to display information
about one thread, and one component (Threads) to contain each Thread
component. However, the goal is to display the comments of a given
thread as well. As such, a third component will be needed, to be called
ThreadDisplayer. It is through this idea that the naming scheme of
Thread and Threads does not suffice. Thus, Thread is now ThreadItem and
Threads is now ThreadList.