Renametimestamp column to be specific to table
Timestamp is a special (but not reserved) keyword in PostgreSQL. Thus, a
conflict might arise if one wanted to cast a date as a timestamp (since
the querying library would not know if one meant timestamp the column or
timestamp the cast). As such, each table that has a timestamp column now
has a timestamp column that is specific to that table (e.g.,
thread_timestamp, comment_timestamp, outfit_timestamp).