Refactor thread query function to be more robust
Previously the query function to retrieve a thread by the date was using
a static value of 86400, that being the number of seconds in a day. It
would subtract the two timestamps and determine if the difference was
less than 86400. However, errors arise depending on the TIME the thread
is posted, as 24 hours after posting will be a different day, but less
than 86400 seconds will have elapsed. Thus, the new query function now
converts the timestampts to PostgreSQL timestamps, and then to dates, so
that we are only comparing dates (e.g. 2019/05/01). The query function
also now has more checks in case the query itself yields 0 results.