Change thread display to be based on date
Previously a user could look at a thread by specifying the thread ID in
the URL. This presents difficulties, as it's tedious to type in a thread
ID, when it's much easier to type in a date. As such, the way to view a
thread on the app is by typing in a date in the URL. This method may be
changed to either include the thread ID as a specifiable URL parameter.
It might also change in the way the user types in the date (and use URI
encoding as well).