Add trimming of query string of URLs and try-catch for is_url_down
The is_url_down function leverages the requests library to make HEAD
requests to check if a given URL is up or down. However, in the event a
URL is passed to this function which the requests library does not know
how to handle (e.g. http://www.ricoh-imaging.co.jp), it will throw an
error and cause the program to exit. So, a try-except block has been
added which allows the program to continue to run in the event invalid
URLs are passed to the function.