Fix JSON processing to process based on type
Up until now, there was a misunderstanding about the way in which the
Imgur endpoint returns data. It was first thought to return different
datatypes based on the type of URL (gallery, single image, or album).
However, this is not the case. The endpoint in fact returns data based
on whether or not any Imgur URL contains more than one image. If it's a
single image, it will return a dictionary. If it has more than one
image, it will return a list of dictionaries. Thus, the way Imgur JSON
is properly handled is by checking the type of JSON itself, and no
longer by checking the type of Imgur URL.