Change Imgur hash extraction to be more precise
Previously a URL was encountered that contained an ampersand in the
hash. This would cause issues, as the Imgur API does not know how to
handle hashes that have an ampersand in the hash. Through testing, it
became apparent that the API can work with the string before the
ampersand, so the extraction of the hash is now achieved by splitting on
an ampersand, which, if none are in the hash, still works as intended.