Update data model to store comments as array instead of object
Previously, comments were stored in an object.

As a bit of background, we initially used
the object model because it lent itself nicely to grouping outfits by
commentID. However, that model adds no additional value beyond the
grouping.

This object had a key:value mapping of
commentID:commentInformation. This works if we're working with 'static'
data, in the sense that the data never changes (both in order and value).
However, if we want to do something like sort the data for easier
viewing, it's somewhat awkward to do so.
This is because we have to use Object.keys and index
into the object via its keys, which is harder to read and wrap one's
mind around (both in a programming sense and in a data model sense).
As a result, we converted the object into an array.