Change saving threads to file from append to write
Previously we had been using the append modifier for opening files. This
was an oversight in retrospect. Specifically, we have a list that has
now been trimmed down to less threads than when the file was first ran.
Thus, we want to overwrite the file with the trimmed list, not append to
it (and thus add/leave duplicates). The way to fix this is to use the
w+ modifier, so we can read the lines and make sure we don't by some
chance add a duplicate, and also so we can otherwise overwrite the
original file.