Remove data visualizations (for now)
At the current stage of this project, I want to focus on "quality of
life"-based tasks and ensure the core functionality of the project is
where I want it. The data visualization was becoming too much of a
hassle to work with, was not visually appealing, and was more of a
nuissance than anything else. So, for now, it has been removed. The base
files are still on the remote repo, but I think I may not use D3 in the
future.